pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "dasproject"
version = "0.0.1"
description = "Das Project"
authors = [
  { name = "Gustav Jacobsen", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
    "dvc>=3.66.1",
    "fastapi==0.115.6",
    "h5py>=3.15.1",
    "isort>=7.0.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "torch==2.6.0",
    "tqdm>=4.67.1",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "optuna>=4.0.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cu124" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
